+++ R/get_language_lookup.R
@@ -61,8 +61,7 @@ extract_field_names <- function(fields_list) {
 #' are contained. Can be obtained with [get_configuration()].
 #' @param language Language short name that is to be extracted, e.g. "en",
 #' defaults to "en"
-#' @param remove_config_names TRUE/FALSE: Should the name of the project be
-#' removed from field names of the configuration? (Default is TRUE.)
+#' @inheritParams get_field_inputtypes
 #'
 #' @returns A data.frame that can serve as a lookup table, with the background
 #' name in the "var" column, and the selected language in the "label" column.
@@ -88,6 +87,9 @@ extract_field_names <- function(fields_list) {
 get_language_lookup <- function(lang_list,
                                 language = "en",
                                 remove_config_names = TRUE) {
+
+  stopifnot(is.logical(remove_config_names))
+
   # if any of the reversed results of grepl are true, we need to skip because
   # the names are not in language-list forma, e.g. "en", "de", "fr"
   # (the reverse/any combination is weird, but we have to reverse the matches