+++ R/onLoad.R
@@ -6,14 +6,19 @@
   options(
     # The list contains a list of options.
     list(idaifield_categories = list(
+      # The list of options contains the default values for the operations.
+      operations = c("Operation", "Trench", "Building",
+                     "Survey", "ExcavationArea"),
+      catalogues = c("TypeCatalog", "Type"),
       # The list of options contains the default values for the categories.
-      layers = c("Layer", "Grave", "Burial", "Architecture", "Floor"),
+      layers = c("Layer", "Grave", "Burial", "Architecture", "Floor", "SurveyUnit"),
       # The list of options contains the default values for the strict categories.
       layers_strict = c("Layer"),
       # The list of options contains the default values for the relations.
       relations = c("isDepictedIn", "isRecordedIn", "liesWithin",
                     "borders", "cuts", "depicts", "fills", "hasInstance",
                     "hasLayer", "hasMapLayer", "isAbove", "isAfter",
+                    "abuts", "isAbuttedBy", "bondsWith",
                     "isBefore", "isBelow", "isContemporaryWith", "isCutBy",
                     "isDepictedIn", "isFilledBy", "isLayerOf", "isMapLayerOf",
                     "isRecordedIn", "isSameAs", "liesWithin")
@@ -23,5 +28,6 @@
 }
 
 .onAttach <- function(lib, pkg) {
-  check_idaifieldr_version(getNamespaceVersion('idaifieldR'))
+  checked_idf <- check_idaifieldr_version(getNamespaceVersion('idaifieldR'))
+  rm(checked_idf)
 }